用户账户添加

+++ dmd-mall/src/main/java/com/dmd/mall/mapper/OmsShippingMapper.java
@@ -28,7 +28,7 @@ public interface OmsShippingMapper extends MyMapper<OmsShipping> {
      *
      * @return the int
      */
-    int deleteByShippingIdUserId(@Param("userId") Long userId, @Param("shippingId") Integer shippingId);
+    int deleteByShippingIdUserId(@Param("userId") Long userId, @Param("shippingId") Long shippingId);
 
     /**
      * Select by shipping id user id omc shipping.
